+++ src/features/orders/components/coversheet/index.ts
@@ -17,4 +17,5 @@
 export { ComplianceHeader } from './ComplianceHeader';
 export { BatchComplianceTable } from './BatchComplianceTable';
 export { DistributedToSection } from './DistributedToSection';
+export { PackageManifestSection } from './PackageManifestSection';
 export { CoversheetActions } from './CoversheetActions';